Re-arrange components using atomic design
