+++ components/atoms/logo.js
