+++ components/organisms/archived-posts.js
@@ -1,13 +1,13 @@
 import { string } from 'prop-types';
 import NextLink from 'next/link';
-import formatDate from '../utils/format-date';
-import Link from '../components/link';
-import Card from '../components/card';
+import formatDate from '../../utils/format-date';
+import Link from '../atoms/link';
+import Card from '../molecules/card';
 
-import { frontMatter as picoCTFWriteup } from '../pages/archive/picoctf-2018-crypto-writeups.mdx';
-import { frontMatter as NibblesWriteup } from '../pages/archive/nibbles-hack-the-box-writeup.mdx';
-import { frontMatter as WhyDymergeSucks } from '../pages/archive/why-dymerge-sucks.mdx';
-import { frontMatter as MrRobotWriteup} from '../pages/archive/mr-robot-1-writeup.mdx';
+import { frontMatter as picoCTFWriteup } from '../../pages/archive/picoctf-2018-crypto-writeups.mdx';
+import { frontMatter as NibblesWriteup } from '../../pages/archive/nibbles-hack-the-box-writeup.mdx';
+import { frontMatter as WhyDymergeSucks } from '../../pages/archive/why-dymerge-sucks.mdx';
+import { frontMatter as MrRobotWriteup} from '../../pages/archive/mr-robot-1-writeup.mdx';
 
 const archivedPosts = [
 	picoCTFWriteup,
