+++ pages/archive/dymergeproject.mdx
@@ -5,8 +5,8 @@ summary: "DyMerge - Dynamic Dictionary Merger"
 ---
 
 import Tweet from 'react-tweet-embed';
-import AsciinemaPlayer from '../../components/asciinema-player';
-import YoutubePlayer from '../../components/youtube-player';
+import AsciinemaPlayer from '../../components/molecules/asciinema-player';
+import YoutubePlayer from '../../components/molecules/youtube-player';
 
 ## DyMerge - Dynamic Dictionary Merger
