+++ pages/archive/kickthemoutproject.mdx
@@ -6,9 +6,9 @@ summary: "KickThemOut - Onion URL Inspector"
 ---
 
 import Tweet from 'react-tweet-embed';
-import FacebookPost from '../../components/facebook-post';
-import AsciinemaPlayer from '../../components/asciinema-player';
-import YoutubePlayer from '../../components/youtube-player';
+import FacebookPost from '../../components/molecules/facebook-post';
+import AsciinemaPlayer from '../../components/molecules/asciinema-player';
+import YoutubePlayer from '../../components/molecules/youtube-player';
 
 ![KickThemOut Logo](/static/images/kickthemoutproject/kickthemout_small.png)
