+++ pages/archive/onioffproject.mdx
@@ -5,7 +5,7 @@ summary: "ONIOFF - Onion URL Inspector"
 ---
 
 import Tweet from 'react-tweet-embed';
-import AsciinemaPlayer from '../../components/asciinema-player';
+import AsciinemaPlayer from '../../components/molecules/asciinema-player';
 
 ## ONIOFF - Onion URL Inspector
