+++ pages/index.js
@@ -1,6 +1,6 @@
-import Page from '../components/page';
-import Hero from '../components/hero';
-import Projects from '../components/projects';
+import Page from '../components/templates/page';
+import Hero from '../components/organisms/hero';
+import Projects from '../components/organisms/projects';
 
 const Home = () => {
 	return (
